função transfereSaldo implementada

+++ src/controle/InterfaceControle.java
@@ -16,7 +16,7 @@ import modelo.Beans.Conta;
 public interface InterfaceControle extends Remote {
     
     public Conta solicitaCriacao(String nome, String senha) throws RemoteException;
-    public int solicitaConsulta(String nome) throws RemoteException;
+    public double solicitaConsulta(String nome) throws RemoteException;
     public ArrayList<String> solicitaExtrato(boolean flag, int id) throws RemoteException;
     public boolean autenticacao(String nome, String senha) throws RemoteException;
     public boolean login(String id, String senha) throws RemoteException;
@@ -24,5 +24,5 @@ public interface InterfaceControle extends Remote {
     //aqui que pega a hora e faz o log da transferencia para ela poder ser consultada por extrato
     //tem que fazer com que o log na conta remetente e destino seja atualizado para a transferencia
     //antes de qualquer outra operacao ser feita
-    public int transfereSaldo(String remetente,String destino,double quantidade) throws RemoteException;
+    public int transfereSaldo(String remetente ,String destino, double quantidade) throws RemoteException;
 }
